requirements.txt
fastapi
uvicorn
ultralytics
numpy
opencv-python
yt-dlp
python-multipart
python-dotenv

